Implement CallList execution
I was hoping to get away from cloning the display list commands ever (in fact by making execute take a command reference we eliminate most of the need for them) but there's a problem with iterating over a display list that's borrowed from self and then borrowing self to execute commands. I can clearly see that this is safe as execute never modifies the display_lists vector, but I don't know of any way to communicate that to the rust compiler.

Luckily, we're not that concerned with performance here, so this clone shouldn't be too terrible.