Successfully hook/patch swap_buffers
Since we don't need to preserve the original functionality, it's fine to just patch the function and not care about what we're overwriting. I've also added code to restore the patched data, which I've confirmed works successfully when the Context is dropped, but I'm not sure we actually need this. Maybe if we end up doing some dynamic patching in the future; who knows.