Swap edge function x/y terms
The purpose of this is to have the edge function be >= 0 in the interior of CCW polygons, rather than <= 0. The pineda paper had them defined for CW triangles.

The difference is subtle, but important, mostly for when we move to fixed point, but it's already relevant - >= 0 is purely a sign bit check in both f32 and i32; <= 0 is a sign bit check AND an actual value comparison with zero. Having an edge function that reduces to ONLY checking sign bits should make the hardware implementation smaller/simpler.