Replace uv/edge functions with wi/orient2d functions
Particularly, we can replace the 'edge function' vars with explicit barycentric weights (barycentric coords up to scale), which we'll need for interpolating values across the triangle anyways.

As-is, the code is equivalent and produces exactly the same results, but now match what's presented in ryg's sw occlusion culling blog series here:

https://fgiesen.wordpress.com/2013/02/17/optimizing-sw-occlusion-culling-index/

This is a highly-recommended series, and not only contains lots of code examples, but goes into great detail about how these examples are derived in a clear, geometric way. So, matching that as closely as possible is a good way to keep the details straight and be able to double-check and debug when things inevitably go a bit haywire. :)