Reduce barycentric weight calcs in bb traversal to adds
The idea here is (and always was going to be!) to calculate the initial weight values for the first traversal point, as well as partitial derivatives for the weights wrt x and y, and then evaluate per-pixel weight values incrementally as running sums over the bb.

Admittedly I'm having some trouble understanding the derivative calculations and how I'm going to calculate those for other quantities, but that's the next step.