Initial texture unpack/test display
This is really disturbing on startup now, but is really useful in order to determine that we're reading data correctly.

Note that this totally ignores the unpacking params specified with glPixelStorei, as well as the internalformat, but this is a great place to cut corners for the purpose of this project.