Initial texturing impl
Totally works! No filtering or mipmaps, the lack of perspective correction is super painful, no multitexturing, and the lack of lighting is _very_ apparent right now. Also we're doing quite a lot of per-pixel work, but I can move that out of the inner loops progressively and do more and more in fixed point. This is awesome!

I'm also noticing a couple coverage issues in the rasterizer it looks like, wonder why that's happening...