Basic bilinear filtering
This shouldn't be applied in all cases and still doesn't mip-map (which I probably won't support for this hw), but looks OK.

A big thing I'm struggling a bit with now is edge cases for texture sampling. For fullscreen overlays in the Mosaik demo, with real GL, there are very subtle wrapping artifacts at the edges of the viewport. In this impl, those artifacts are much more pronounced. I'm not sure why this is; the offset we use to sample barycentric coordinates at pixel centers should be good, and the offsets we then subtract when converting the weighted texture coords to texel centers should also be correct. Maybe it's a precision issue? Maybe the actual GL impl is using different mip levels for rendering these overlays? I'm not entirely sure, but it's a bit frustrating.