Remove explicit DllMain
DllMain is actually pretty restrictive and it's best if we don't use it at all, as long as we can defer any required initialization until later. In our case, we can always just lazily create the Context whenever it's requested, so it can be deferred until the first implemented entry point call, which is fine, and less code anyways.