Initial perspective correction for s, t coords
Ended up writing it out longform, which sucks, but at least it's explicit, which I like (for now). I also dropped the r and q coords from tex coords for the entire pipeline; none of the test demos use it, and I don't plan on implementing it.

Note that I didn't do any perspective correction for z/depth; the GL 1.2.1 spec actually mandates that depth be interpolated linearly (and suggests color can be as well). This is somewhat interesting; I would have expected differently, and I'm not entirely sure why this is OK. However, I'm happy to accept the optimization.