Convert all interpolants to fixed point
Made a bit of a mess here, but hey, it appears to work quite well. It's likely that not all of these are the final bit counts (for example, s and t seem to be suboptimal in some cases), but that should be fairly easy to tweak.

I also fixed a problem with culling, where dividing edge function values by scaled_area (to scale barycentric coordinates properly) would actually flip backwards edges, meaning that we were always drawing both back and front facing polygons all the time! This is now fixed for the actual interpolated edge function values (which are now not scaled, which has the nice side-effect that it keeps their dynamic range more predictable.. or maybe not?) and backfaces are explicitly detected/fixed (still need to use proper culling state).

Note that many of these params are converted back to floats in the per-pixel code currently; this is only for convenience, and is OK for now, as it shows that there are still enough bits to interpolate properly (in most cases).

Note also that this makes the imperfect coverage/lack of proper fill rule issues more apparent. It's not as bad as I feared (since I was able to use a fair number of fractional bits for edge functions), so I think I'll keep that on the back burner for now, but something to consider still.