Reduce resolution, start on fixed point color math
Not entirely there yet, but getting closer. I also started imposing a 128x128 max texture size, which is what I plan to have for the hw.

One fun thing here is that since the test demos expect to run in 640x480 and set the viewport size accordingly, we can just half the glViewport args to force the projection to fit the test window.