Propagate/use vertex color in rasterizer
This isn't properly interpolated still and it's the same hack as before, but this helps me keep track of what regressed when I switched to tile rendering. At the very least we should try to make things match as best we can in order to catch bugs while massaging stuff.